28/03/2025

+++ KOP/KOP.Common/DTOs/GradeDTOs/ProjectDTO.cs
@@ -10,8 +10,8 @@
         public DateOnly StartDate => DateOnly.FromDateTime(StartDateTime);
         public DateTime EndDateTime { get; set; } = DateTime.Today;
         public DateOnly EndDate => DateOnly.FromDateTime(EndDateTime);
-        public int SuccessRate { get; set; }
-        public int AverageKpi { get; set; }
+        public double SuccessRate { get; set; }
+        public double AverageKpi { get; set; }
         public double SP { get; set; }
     }
 }
